time_nml is meant to be used at the model Top level and read the calendar, start and end time and the restart_time.
the time_config will be used later in the subsequent model type to use the date_to_time functions to provide the parameter to the model. This seems to be different from Renes Approach therefore we have to discuss it again.

The run_nml is cleaned up anyway fromthe init/end time stuff - but not from the dtime, run_time values!