Further work on dummy coupled models (Feature #1358)
The master namelist has changed. I apologize for this, but it is necessary
for integrating multiple coupled dummy models. A technique devised by Rene
allows reading multiple instances of a namelist, thus allowing the
definition of multiple components.

If you use the automated script generator, the correct master namelist is
going to be created for you. If you have manual runscripts, then change
the master namelist according to the following example (only one model
should be present). Again, I apologize for the inconvenience.


&master_nml
 lrestart               = .false.
/
&time_nml
 ini_datetime_string = "2008-09-01T00:00:00Z"
 dt_restart          = 5184000
/
&master_model_nml
  model_type=1
  model_name="ATMO"
  model_namelist_filename="NAMELIST_test_hat_jww-moist_cld-cnv-vdf"
  model_restart_info_filename=""
  model_min_rank=1
  model_max_rank=65536
  model_inc_rank=1
/
&master_model_nml
  model_type=2
  model_name="OCEAN"
  model_namelist_filename="NAMELIST_test_hom_lsm"
  model_restart_info_filename=""
  model_min_rank=1
  model_max_rank=65536
  model_inc_rank=1
/