Improvement of idex search and minor correction
The index search for the aquaplanet coupling is now more flexible
and able to handle an arbitrary ordering of the global vertex idices.

All coupler debug output is now behind an Fortran if and deactivated
per default. The debug output unit is now determined by an INQUIRE
loop over units.