[trunk/icon-dev]
1.- Solved the problem  with duplicate_prog_state. The problem 
was that prog%tracer and prog%tke are not allocated in some cases. 
Now  they are initialized to  NULL
Then (in duplicate_prog_state) they are duplicated only if they are associated.
For any reason in the SX9 the problem did not show up without mpi. 

2.- Added OpenMP directives to  
testcases/mo_nh_mrw_exp.f90