[trunk/icon-dev]
- Instantaneous values of latent heat flux and sensible 
heat flux had not the write sign in the output in case of 
Raschendorfer turbulence. It is corrected now. The 
average/accumulated values had already the correect sign 
in the output.

- Implemented the output of instantaneous and average/accumulated 
latent and sensible heat fluxes for ECHAM turbulence in ICONAM. 
It works for iwtr=1 and nsfc_type = 1 (this is the setup of the APE) and 
for iice=1 and nsfc_type = 1. For other more real cases it is not implemented.
The subroutine surface_fluxes calculates not only the instantaneous fluxes, 
but also the accumulated fluxes. For ICONAM, only the instantaneous values 
are passed to vdiff, later the accumulated or averaged values are calculated 
(this part was already implemented). It is not the most efficient way, but 
it was faster to implement. Any case, I understood we do not plan to use vdiff
for ICONAM  in the future.