[trunk/icon-dev]
More work with the Weisman Klemp test case for the limited area model.
Seems to be well initialized and run at least for 2 hours (the run time 
in the Weisman Klemp paper experiments).
I run it with microphysics, 
turbulence and saturation adjustment as physical parametrizations. 
For this test case there should not be fluxes exchange at the surface. 
So each time step, before the turbulence runs, the potential temperature 
at surface is set to its value at the lowest level as a first solution.
First results give smaller vertical velocities than the reported by 
Weisman Klemp.