Changed the way of importing patches for saving memory in parallel runs
Now only a so called "basic patch" is allocated for the full domain
on all processors and read in before subdivision.
This basic patch contains only the arrays needed for subdivision.
After subdivsion the remaining arrays of the patch are allocated
(for the distribuited domain only) and read from input file
or calculated from other variables.