Implementation of new JSBACH.
The new JSBACH has its own repository at https://svn.zmaw.de/svn/jsbach and is pulled in using 
the svn:externals property on directory src/lnd_phy_jsbach. Apart from that, this commit includes
the changes necessary to make icon run with this JSBACH.

This is a technical commit for testing. At the moment, Thomas' preliminary implementation is still included and 
actually provides the surface conditions for the atmosphere. The new JSBACH just runs in parallel
and generates its own output files. Once the new JSBACH is complete (in a JSBACH-lite version), it
will be fully switched on and Thomas' implementation will be removed.