Changed default in AMIP run script exp.r2b4_amip_thunder.run to not do a restarted run.
Since the previous commit changed the growing time of topography in icoham to 1 day, it's ok to
initialize a new experiment. Restarts from an old restart file don't work if the namelist
attributes in the restart file contain a parameter that has been deleted from the code since
creation of the restart file.