implement spot values for ocean output (refs #3334)
this is a very simple implementaion with used additional memory. the
pointer notation just does _not_ work, i.e. the code is executed without
errors but the spot value output is still wrong. I hope I can go back to
the pointer version in the future, but now I just need the spot values
to be correct.