Bug fix for the growth of mountains in the initial days of the AMIP experiment.
Background:
- The geopotential is needed for the dynamics, in order to compute the geopotential gradient force on the terrain-following model levels. The geopotential is computed using the heigth of the surface and integrating the hydrostatic equation to each model level. This is done by calling the subroutine "update_tempv_geopot" without the optional argument "opt_lgeop_wrt_sfc".
- The geopotential WITH RESPECT TO THE SURFACE is needed for the computation of the vertical diffusion effects on temperature and the surface flux of sensible heat. This is computed as the geopotential for a surface elevation set to zero. This is done by calling the subroutine "update_tempv_geopot" with the optional argument "opt_lgeop_wrt_sfc" set to ".TRUE.".

Problem:
The code added to let the surface grow linearly from zero to the full height was inserted for the case "opt_lgeop_wrt_sfc=.TRUE." and "lamip=.TRUE" had the effect that the vertical diffusion was given the full geopotential instead of the geopotential above ground.
Through a side effect the dynamics still got to know about the growing surface, although the code section called by the dynamics to get the geopotential missed the code to let the surface grow.

Change:
Now vertical diffusion get the geopotential above the surface as input, and the surface is grown at the place where the full geopotential needs to be computed.

Tests:
Currently only compilation is tested for GCC. An experiment will be started once the machines are available again.