Until 2013-06-14 the Fortran template files "template_xyz.f90" were available in src/templates. Then they were moved out of the way to doc/templates in order to avoid complications in the Makefile construction due to the ending ".f90" of the template files. But the src/templates directory remained. This had the effect that the templates were no longer found.
Here the template files are moved back to the old directory, where they are better placed. The file names are changed from "template_xyz.f90" to "xyz.f90_template" so that they are not compiled together with the other Fortran files of the model. "template_main.f90" is changed to "program.f90_template" as it is about Fortran PROGRAM units. Thus "xyz.f90_template" files are about Fortran XYZ units.