Correction for the conversion factor for heating rates in the case of constant volumes (i.e. the non-hydrostatic setup).
The error was in line 1462 in the moisture correction term, which was computed using the "vtmpc2" constant that, however, is valid for constant pressure conditions.

The new line uses directly cvd and cvv instead of auxiliary constants.

The equivalent (and correct) line for constant pressure conditions (line 1465) is rewritten for clarity.