This replaces the cloud cover, cloud microphysics and convection parameterizations of a pre-ECHAM6 version by the parameterizations of ECHAM6.2.
This work started with r14502. The replacement was carried out and tested in a AMIP like simulation by Thorsten Mauritsen.

icon-mau-phys@r16326 was tested with the entire Buildbot test suite. This showed the known issues of trunk/icon-dev. In addition test_couple_atmo-ocean failed in a strange way, possibly for external reasons. In a 2nd trial also this test succeeded and additional tests on this branch for this experiment alone for r16313, r16317, and r16319 were successful.

Source code changes apply only in src/atm_phy_echam, src/namelists and src/configure_model. Additional changes apply to experiment descriptor files. None of these changes interferes with NWP or ocean configurations.