echam physics: Merge of r3673 from echam-dev (exchange coefficients)
Requires a new JSBACH revision (>= r81) because of removal of rel_hum_srf variable (zhsoil in echam).

echam-dev, r3673:
zhsoil is replaced by zcsat, zcair in the calculation of the exchange coeff.
to make it more consistent with the surface energy balance formulation (see
also rev. 7222 of cosmos-landveg).
zhsoil is removed completely.