[icon-aes-mag] Bug fix for updating the mass mixing ratio with respect to moist air, cf.  r29114.
As for r29114, the test exp.atm_amip_test modified for more output is used to check the mass changes. Differences from step 4 to step 5 of the global mean vertical integral of air mass, total water mass, and dry air mass in kg/m2 are as follows: 

ldrymoist=.FALSE.:
* air: +1.0732e-10 kg/m2 (air=rho*dz)
* h2o: +0.0068988 kg/m2 (h2o=vap+liq+ice)
* dry: -0.0068988 kg/m2 (diagnosed as air-h2o)

ldrymoist=.TRUE. :
* air: +0.0061497 kg/m2 (air=dry+h2o after physics)
* h2o: +0.0061497 kg/m2 (has sources and sinks in physics)
* dry: +1.4552e-11 kg/m2 (=air-h2o before physics)

Considering that the base values of the mean vertical integrals of air and dry air are 10045 and 10022 kg/m2, respectively, the relative error is of order 1.e-14 to 1.e-15, which is the expected numerical accuracy. So global mass is conserved to the numerical limit, either for air mass (ldrymoist=.FALSE.) or for dry air mass (ldrymoist=.TRUE.).

Now the effect on the simulated climate needs to be explored.