Merge icon-les.git:icon-les/icon-les-dev (SHA: 4ae80b97) into icon.git:master.
1. Replace JSBACH macro expansion based on cpp include file with external python pre-processing
   script "dsl4jsb.py" (cpp expansion didn't work on DWD Cray).
2. It is now possible to use platform-specific option settings in config/config-options.
3. Bug fix in mo_var_list:add_var_list_element_5d for the case that an already allocated pointer
   p5_* is passed.
4. JSBACH now contains the code for the standalone version, although currently this works only
   together with ECHAM.
5. JSBACH now computes and returns to the atmosphere the heating by melting of snow on the canopy
   instead of the temperature tendency.

Buildbot is green.
The update test of checksuite.icon-dev is not passed, but there is only a small difference in the
lowest-level temperature tendency in the order of E-20 which is due to the changed computation of
the heating by melting of snow on canopy.
The nproma test on an MPI workstation fails with one very small difference on the order of E-300,
but this is already the case in the previous version before this commit.