[icon:icon-mpiesm2-dev] Update of JSBACH to jsbach:develop@8c0ad871 and change in exp.atm_amip experiment descriptor file
New namelist switch "l_compat401" to configure JSBACH in such a way that results of AMIP experiment are similar to icon:icon-2.1.00 resp. icon-aes:icon-aes-1.2.00 (jsbach-4.01). This switch is set to .true. in the exp.atm_amip experiment descriptor file.

In addition, changes to the multi-layer snow model in JSBACH if this switch is .false.

Squashed commit of the following:

commit 8c0ad87167b3a02f625924e8a34988779b75eb1f
Author: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Date:   Sun May 21 19:58:15 2017 +0200

    [icon-les:icon-les-dev] Update of experiment descriptor files for new jsbach switches

commit 3018eefb3a92a361db70732125411e0ba741b060
Author: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Date:   Sun May 21 19:43:18 2017 +0200

    [icon-les:icon-les-dev] Update JSBACH to jsbach:develop@82f6b1ea