[icon-nwp-dev] remove PROTECTED attribute from variable time_config
Cray compiler v8.5.8 issued an error where time_config is used as the target in a pointer assignment statement.

F2008 standard states: A non-pointer object that has the PROTECTED attribute and is accessed by use
association can not appear in a variable definition or as the target in a pointer assignment statement.