[icon-aes:master] [icon-aes:icon-aes-dbg3] Fix for surface saturated specific humidity over lakes (water and ice)
The values computed for qsat_lwtr and qsat_lice for the new time step are now returned to update_surface and put into the bb_btm matrix, so that they are used in the computation of new atmospheric moisture. Not doing so leads to a water imbalance.

This commit uses JSBACH version jsbach:jsbach4icon-aes@bfa6236 (jsbach-4.20 plus this fix).