[icon-aes:master] Squashed commit of [icon-aes:icon-aes-nesting-phy], #09916588
Implementation of the changes developed by Vera Maurer for nesting in
the global AMIP simulation:
- The changes for the parameterizations were manually merged based on
icon-aes:icon-aes-nesting 22965f09. The manual merge was needed because
of the complications of the merge history on icon-aes-nesting, which did
not allow a simple differentiation between the changes in different
parameterizations.
- Additional developments allow the handling of external files for
multiple grids, as needed for nesting, for: SST/SIC, tropospheric Kinne
aerosols, and ozone. Here the existing single grid variables have been
replaced by arrays of specific variable types.
- Buildbot tests pass/fail as in the base version icon-aes:master
- Scripts for nested experiments are not contained, because no tests
have been conducted so far.

Remarks:
Gustavo's version of PSrad requires all input stored on indices
1:jce-jcs+1 in the cell-dimension. The shift of all input arrays and the
reverse shift of the output arguments is implemented in
mo_psrad_interface, and realized in the subroutine
shift_and_call_psrad_interface_onBlock that is called instead of
psrad_interface_onBlock if jcs>1. This shift, however, misses the
aerosol input arrays, because these are not included in the argument
list. Therefore an ad-hoc shift operation is included in the subroutine
set_bc_aeropt_kinne of mo_bc_aeropt_kinne that is called inside
psrad_interface_onBlock to interpolate the external monthly data, stored
in jcs:jce, to the current time.