[icon-aes:master] Squashed commit of icon-aes-phymem, #ae557c32.
The memory usage of by the ECHAM physics is reduced by allocating global
memory for several variables only if needed. This concerns the variables
for tendencies of specific parameterizations and a number of diagnostic
variables.
- Tendency variables from processes are allocated if they need to be
stored for usage in later time steps. This is the case if dt_prc > dt_dyn.
- Further, tendency variables and a number of diagnostic variables are
allocated as global variables if they are requested for the output,
which is diagnosed by the "is_variable_in_output" function.
- Within the interface subroutines "interface_echam_prc" it is tested if
pointers for global variables have been associated with memory. If this
is the case then local variables are copied to or retrieved from the
corresponding global variables.

This technique has been applied to many, but not all variables where
this would be possible. The focus was on the tendency variables and
other 3d-variables.

Results of atm_amip_test changed with GCC for numerical reasons, due to
changes in interface_echam_rht, where d=c+b+a is replace by z=b+a;d=c+z.
(Not tested with other compilers).

The model passed nproma, mpi, and restart tests on a desktop PC at MPI-Met.
Results are changed as for GCC compiler.