Squashed commit of [icon-aes:icon-aes-cover]:
The cloud cover scheme for ECHAM physics is modified as follows:

1. The control parameters for the diagnostic cloud cover is separated
from the cloud physics control parameters, and now organized in the
configuration state echam_cov_config, which can be controlled via the
namelist echam_cov_nml.

2. A 0/1 cloud cover scheme is introduced as an alternative to the
fractional cloud cover scheme. Depending on the switch icov of
echam_cov_config, either the fractional (icov=1=default) or the 0/1
scheme (icov=2) is used.
The 0/1 scheme diagnoses 0 or 1 (m2/m2) cloud cover if the cell mean
relative humidity is < or >= csat, respectively. The default value of
csat is 1.csat can be modified through the namelist echam_cov_nml.

3. Control parameters for vertical levels are replaced by control
parameters in height, from which the level indices are now diagnosed.
Thus it is no longer necessary to change level indices of the cloud
cover parameterization if the vertical grid is changed, unless the
heights must be changed for a physical reason. The following control
parameters are changed:
- echam_cld_config%jks   =15 --> echam_cov_config%zmaxcov
- echam_cld_config%jbmin =40 --> echam_cov_config%zmaxinv = 2000m
- echam_cld_config%jbmax =45 --> echam_cov_config%zmininv =  200m

4. In the same way the control parameter jks for vertical levels in
the cloud microphysics is changes to a control parameter in height:
- echam_cld_config%jks   =15 --> echam_cov_config%zmaxcld

5. A new control parameter zmaxcloudy is introduced in the echam
physics configuration, to be used as a specifier for the top height of
cloud processes:
- echam_phy_config%zmaxcloudy =33000m
This parameter is used to initialize zmaxcov and zmaxcld:
- echam_cov_config%zmaxcov = echam_phy_config%zmaxcloudy
- echam_cld_config%zmaxcov = echam_phy_config%zmaxcloudy
An equivalent control parameter for convection does not exist.

The control parameters for heights have default values which result in
vertical level indices, which are equal to the previously directly
specified indices. Therefore the results of the exp.atm_amip_test do not
change.

The 0/1 scheme was compared to the fractional scheme in an AMIP
experiment over one year. It behaved reasonable, meaning that the time
mean cloud cover was generally similar to that of the fractional scheme,
except that the contrast between cloudy regions (mid latitudes) and less
cloudy regions (subtropics) tend to be amplified.

In tests on a Mac, using "./configure --with-fortran=gcc --with-mpi
--with-openmp --with-eccodes=/opt/local --enable-mixed", the experiment
exp.atm_amip_test passed all tests (base restart nproma mpi omp).