[icon-aes:master] Squashed commit of #d12dc89 [icon-aes-slaminit]
To enable a proper initialization of an aqua planet experiment on a
planet with scaled Earth radius, additional control parameters are
implemented in grid_nml and nh_testcases_nml:

(1) For the Jablonowski-Williamson initial state the parameters jw_u0 and
jw_temp0 have been added to the namelist nh_testcases_nml.
 - jw_u0    = maximum zonal wind,
 - jw_temp0 = horizontal average temperature at surface

(2) Two switches lrescale_timestep and lrescale_ang_vel are introduced
in the namelist grid_nml to control whether the grid rescale factor for
a "small planet" should be used to scale also the timestep and the grid
angular velocity.  The default values are:
- lrescale_timestep = .FALSE.
- lrescale_ang_vel  = .FALSE.
Existing scripts, using grid_rescale_factor different from 1 have been
adjusted by adding lrescale_timestep = lrescale_ang_vel  = .TRUE. so that
these experiments are not changed.

Further, in mo_turbulence_diag.f90 the earth_angular_velocity is replaced
by grid_angular_velocity to make it consistent with the rest of the
model if the angular velocity is changed.