[icon-les:icon-les-dev] Bug fix for restartability related to cloud microphysics (mo_echam_phy_init)
This bug fix, which adds a missing differentiation between echam cloud microphysics and graupel cloud microphysics, solves a failing restart test. The problem consisted in setting rsfl and ssfl to zero after restarting. This must be done if a restart is performed with cloud microphysics switched off. The test for this operation included only the usage of echam cloud microphysics. Now that there exists the alternative graupel cloud microphysics, rsfl and ssfl must be set to zero after a restart only if both schemes are switched off.