[icon/icon-aes:master] [icon-aes-pressurevars] Bug fix: init_echam_phy_params must be called earlier
Put the call of init_echam_phy_params back to its previous position, which was
changed in icon-aes:icon-aes-pressurevars:8b720f63.

The reason is that the initialization of the tracer indices and the related
control indices for tracer index ranges must happen before the setup of the
tracer fields in mo_nonhydro_state and mo_echam_phy_memory, if more than the
basic three tracers for water vapor, cloud water and cloud ice are used.

Thus this correction is not needed for exp.atm_amip and the related tests,
but for experiments using "graupel" microphysics and similar.