[icon-les:master] Changes in time control of JSBACH
New JSBACH version: jsbach:master@7b300cbc

o The time control must consider the dates and times and events for each model_id,
  i.e. domain, individually. Previously, events like is_newday or is_newyear in
  JSBACH were computed wrongly for the nested domains.
o Also, previously is_newday or is_newyear was true when the new time step was at
  midnight of a day. This contradicts the logic in atm_phy_echam ... instead, these
  events are now true when the new time step is at one step AFTER midnight (i.e.
  mtime_old is at midnight). mtime_old at midnight means that the old day has been
  completely simulated, and only then should e.g. sums of the previous day be computed.
o Finally, is_time_ltrig_rad_m1 was previously wrong, i.e. albedo was updated at the
  radiation time step (and therefore after radiation was called), instead of at one
  time step before the radiation is called.

Because of the last two points, results are changed in atm_amip experiments.

Note: Although this fix works for nested domains, commit icon.git:icon-2.6.0-rc@67973d12
contains changes that make the model crash currently with nested domains. Unfortunately,
this commit has many changes (many related to the GPU port, but not all) and needs to
be debugged for nested domains to work with new checkouts of branches that are
up-to-date with icon-2.6.0-rc+ . (2021-07-28: not sure if this is still valid)